updated the formatting of the code and created a new file RidesTemplate which follows the DRY principle better, may delete IOARides.js and USFRides.js. Also allowed user to sort
