+++ frontend/src/components/RidesTemplate.js
@@ -1,13 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import './css/landing.css'
-import RidesTemplate from './RidesTemplate';
 
-
-//////////////////////////////////////////////////////////////////////////////////
-// THIS FILE MAY BE DELETED LATER, MOST OF THIS CODE CAN BE FOUND AT /RidesTemplate.js
-///////////////////////////////////////////////////////////////////////////////////
-
-function IOARides()
+function RidesTemplate(rides)
 {
     const [rideContent, setRideContent] = useState([]);
 
@@ -16,32 +9,7 @@ function IOARides()
     const alphabetical = 0;
     const waitTime = 1;
 
-    var sortBy;
-
-    var rides = [
-        ["Camp Jurassic", open],
-        ["Caro-Suess-el", 5],
-        ["Doctor Doom's Fearfall", 15],
-        ["Dudley Do-Right's Ripsaw Falls", 45],
-        ["Flight of the Hippogriff", 45],
-        ["Hagrid's Magical Creatures Motorbike Adventure", 90],
-        ["Harry Potter and the Forbidden Journey", 35],
-        ["Hogwarts Express - Hogsmeade Station", 10],
-        ["If I Ran The Zoo", open],
-        ["Jurassic Park Discovery Center", open],
-        ["Jurassic Park River Adventure", closed],
-        ["Jurassic World VelociCoaster", 100],
-        ["Me Ship, The Olive", open],
-        ["One Fish, Two Fish, Red Fish, Blue Fish", 15],
-        ["Popeye & Bluto's Bilge-Rat Barges", closed],
-        ["Pteranodon Flyers", 35],
-        ["Skull Island: Reign of Kong", closed],
-        ["Storm Force Accelatron", 5],
-        ["The Amazing Adventures of Spider-Man", 50],
-        ["The Cat in The Hat", 5],
-        ["The High in the Sky Seuss Trolley Train Ride!", 5],
-        ["The Incredible Hulk Coaster", 45]
-      ];
+    var sortBy = alphabetical;
 
     function sortRides()
     {
@@ -81,22 +49,21 @@ function IOARides()
         }
         setRideContent(rideButtons)
     }
-    /*useEffect(() => 
+    useEffect(() => 
     {
         createRideButtons();
-    }, []);*/
+    }, []);
 
     return(
-        /*<div id='rides'>
+        <div id='rides'>
             <label for="sortBy">Sort by:</label>
             <select id="sortBy" name="sortBy" onChange={() => sortRides()}>
                 <option value={alphabetical}>Alphabetical</option>
                 <option value={waitTime}>Wait Time</option>
             </select>
             {rideContent}
-        </div>*/
-        RidesTemplate(rides)
+        </div>
     )
 }
 
-export default IOARides;
+export default RidesTemplate;
