+++ frontend/src/components/USFRides.js
@@ -1,5 +1,10 @@
 import React, { useState, useEffect } from 'react';
 
+
+//////////////////////////////////////////////////////////////////////////////////
+// THIS FILE MAY BE DELETED LATER, MOST OF THIS CODE CAN BE FOUND AT /RidesTemplate.js
+///////////////////////////////////////////////////////////////////////////////////
+
 function USFRides()
 {
     var rideNames = ["Illumination Theater",
