+++ frontend/src/components/css/landing.css
@@ -26,7 +26,7 @@
     right: 0px;
 }
 
-.rides {
+#rides {
     margin-top: 200px;
     display: flex;
     align-items: center;
